requirements.txt
nextcord==3.0.1
python-dotenv==1.0.1
gamercon-async==1.0.7
aiohttp==3.10.11
aiosqlite==0.20.0
pytz==2024.2
steam==1.4.4